// src/constants/constant.enum.ts
export enum ResponseMessageEnum {
  SUCCESS_RECEIVED = "Success",
  SOMETHING_BAD = "Something bad.",
  INVALID_CREDENTIALS = "Invalid login credentials.",
  INVALID_CART = "Cart Not Found",
  INVALID_TOKEN = "Invalid TOKEN",
  TOKEN_EXPIRED = "expired",
  TOKEN_VERIFIED = "Token verified !",
  OTP_SENT = "OTP sent successfully.",
  OTP_SENT_FAIL = "Fail to generate OTP.",
  EMAIL_VERIFICATION_SUBJECT = "Cryptopill: Please verify your Account",
  EMAIL_VERIFICATION_MESSAGE = "Link to Portal for Verification",
  INVALID_PASSWORD = "Old password is Invalid",
  CONFIRM_PASSWORD_INCORRECT = "Please make sure you new password matches the confirmed one",
  SUCCESS_PASSWORD_CHANGED = "Password Changed Successfully",
  ADD_DELEGATES_SUCCESS = "Added Delegates Successfully",
  STAFF_ASSIGNEE_TO_OTHER_WORKPLACE = "Staff Already Assigned to Different Workplace",
  UPDATE_DELEGATES_SUCCESS = "Update Delegates Successfully",
  REMOVE_DELEGATES_SUCCESS = "Removed Delegates Successfully",
  DELEGATE_NOT_EXITS = "Delegates does not exists",
  PATIENT_SEARCH_DOCTORS_FAILURE = "No Nearby Doctors Found",
  SCHEDULE_NOT_ASSIGNED = "No working schedule found",
  DEFAULT_RATING = 3.5,
  INVALID_DOC_SEARCH_CASE_FILTER = "Cannot filter on basis of both Specialty & HealthConcerns Please select either One",
  BLOCK_SCHEDULE_FAILURE = "The given schedules are already blocked",
  NO_SCHEDULE_FOUND = "No working schedule found",
  NO_SLOTS_FOUND = "The given slots doesn't exists",
  INVALID_SCHEDULE_TYPE = "Please provide slots for the given type",
  NONEWWORKPLACEFOUND = "No Workplaces Found",
  USER_NOT_EXISTS = "User entry does not exists",
  USER_ALREADY_EXISTS = "User already exists",
  USER_VERIFIED = "User is Verfied",
  NO_SUCH_DELEGATES = "No Such Delegates entry found",
  BOTH_DOC_PATIENT_NOT_ALLOWED = "Cannot filter on basis of both doctor & patient select either one",
  UNVERIFIED_NEW_MESSAGE = "Admin side verification is pending.",
  UNVERIFIED_EDIT_MESSAGE = "Admin side verification for modified fields is pending.",
  DECLINED_MESSAGE = "Admin side account verification failed.",
  VERIFIED_MESSAGE = "Verified by admin.",
  INVALID_BOOKING_ID = "Booking Id is not Valid.",
  INVALID_DR_WORKPLACE = "Workplace ID added for Doctor is not Valid.",
  DOCTOR_ONLY_GIVE_PRESCRIPTION = "Invalid Token Only Doctor can Add a Prescription for Patient",
  PRESCRIPTION_ADDED_SUCCESS = "Prescription Added Successfully",
  MEDICINE_NOT_FOUND = "Medicine Name Not Found in Database",
  INVALID_PRESCRIPTION = "Prescription Id is invalid",
  PHARMACY_ACCEPT_REQ_ORDER = "Request Order Accepted Successfully for Pharmacy",
  LAB_ACCEPT_REQ_ORDER = "Request Order Accepted Successfully for Laboratory",
  NO_NEARBY_PHARMACY = "No pharmacies found in your area. Please expand your search radius.",
  NO_NEARBY_LAB = "No labs found in your area. Please expand your search radius.",
  SIGNUP_CONFIRM_PASSWORD_INCORRECT = "Please make sure you password matches the confirmed one",
  RATING_SUCCESS = "Rating & Review Submitted",
  DOCTOR_SCHEDULE_CHANGE = "Doctor no longer working on this schedule",
  OTP_RESEND = "OTP re-sent succesfully"
}